requirements.txt
textual>=0.24.1
huggingface_hub>=0.16.0
datasets>=2.12.0
nbformat>=5.8.0
nbconvert>=7.2.9
gitpython>=3.1.31
tqdm>=4.65.0
python-dotenv>=1.0.0

